Add Makefile for creating collectd-mlab packages.
The collectd-mlab Makefile is the primary entry point for building the
collectd-mlab packages. The Makefile runs style and unit tests before
creating three RPMs from collectd-mlab.spec. If the style or unit tests
fail, no RPMs are created. But, the Makefile does include a target for
only the rpms, so 'make rpm' works regardless of whether there are
passing tests.