Move logging for missing vserver names.
When a found xid is not listed in the known set of vserver names, a
current list is requested. However, if it is still not found, this is an
error because an xid is active in the system that has no name. So, this
patch logs this situation as an error.