Change FakeVsysBackend to have a thread instance.
This patch changes FakeVsysBackend to have a thread instance rather than
derive from a thread. This adds two new methods to FakeVsysBackend
interface to 'start' and 'shutdown' the backend, rather than relying on
the inherited threading.Thread methods. This also updates tests to use
the new interface for FakeVsysBackend.shutdown.