Remove LastTime class and separate locking & timestamps.
This change removes the LastTime class. This class fulfilled two needs:
1. provide a file-based lock to guarantee only one process runs at a time.
2. record the last export time to use as the next run's start time.

This change provides the above needs using separate mechanisms.

Locking is provided by a basic LockFile interface. When a process begins it
acquires the lock or exits. When the proces is done, the lock is released. The
lock file is separate from the "last export time" file.

Timestamps are still recorded in the mtime of the "last export time" file. Now,
the mtime is read using "get_mtime" and set using "update_mtime".