Add a collectd watchdog and corrupt RRD cleanup.
Collectd will crash if RRD files become corrupt. This change adds a
watchdog script that will run every five minutes and check if collectd
is running. If it is not, then an integrity check is performed on all
RRD files and collectd is restarted.

This change also adds the watchdog scripts to the collectd-mlab package.