Add disco python site-package & network module. (#32)
Add disco python site-package modules: network, arp, and route.

This change introduces a new directory and python package convention for
mlab modules. The "mlab.disco" package will be installed in a system
site-packages directory so that command line tools may import it easily.

Three modules are included:

arp - for reading an ARP cache.
route - for reading the network routing table (until python-netifaces is upgradeable to 0.10+)
network - which uses the arp and route modules to lookup local and uplink MAC addresses.