Return VLAN id 1 for cisco auto discovery. (#56)
Previously we expected that a Cisco switch would be configured either
with no VLANs (defaulting to vlan id 1) or with at least one vlan (not
1). In practice we dicovered that one site included a configuration
where M-Lab nodes were running on vlan 1 but a second vlan was also
configured unrelated to the M-Lab nodes. This caused the auto discovery
to fail.

To This change returns vlan id 1 and allows such configurations to
succeed.